# crates/indexer/deepbook-indexer-openapi.yaml
openapi: 3.0.3
info:
  title: DeepBookV3 Indexer API
  version: "1.0.0"
  description: |
    REST endpoints for DeepBookV3 order book + analytics data.
    Volumes are returned in the smallest unit of the asset unless stated otherwise.
servers:
  - url: https://deepbook-indexer.mainnet.mystenlabs.com
    description: Public DeepBookV3 Indexer
...
paths:
...
  /pyth/shims/tradingview/history:
    get:
      tags: [Pyth]
      summary: Pyth Pro OHLC chart history
      description: |
        TradingView-compatible chart history backed by the authenticated Pyth
        Pro History API. Successful responses are cached by canonical symbol,
        resolution, and minute-aligned bounds.
      operationId: getPythChartHistory
      parameters:
        - in: query
          name: symbol
          required: true
          schema: { type: string }
          description: Server-configured Pyth Pro symbol, such as Crypto.BTC/USD.
        - in: query
          name: resolution
          required: true
          schema:
            type: string
            enum: ["1", "2", "5", "15", "30", "60", "120", "240", "360", "720", D, W, M]
          description: TradingView candle resolution.
        - in: query
          name: from
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
          description: Inclusive start time in Unix seconds.
        - in: query
          name: to
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
          description: Inclusive end time in Unix seconds.
      responses:
        "200":
          description: TradingView-compatible OHLC arrays.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PythTradingViewHistory"
        "400":
          description: Invalid query or symbol is not configured.
        "429":
          description: Pyth Pro rate limit response.
        "502":
          description: Pyth Pro could not be reached or returned an invalid response.
        "503":
          description: Pyth Pro is not configured.
...
components:
  schemas:
...
    PythTradingViewHistory:
      type: object
      required: [s, t, o, h, l, c, v]
      properties:
        s:
          type: string
          enum: [ok]
        t:
          type: array
          items: { type: integer, format: int64 }
        o:
          type: array
          items: { type: number, format: double }
        h:
          type: array
          items: { type: number, format: double }
        l:
          type: array
          items: { type: number, format: double }
        c:
          type: array
          items: { type: number, format: double }
        v:
          type: array
          items: { type: number, format: double }
...